Added record editing
Records can now be edited, addRecordController has been reused for the most part, but title is changed and fields are set automatically with the case number locked. Added refresh button to the table view. Known issue with edited records not updating in the table until another record is deleted or added, will put it on trello.