Implemented switching between datasets from the UI
Added isEmpty checker to CrimeRecordManager for the case a user is importing for the first time on a dataset, so they don't have to specify importing to current dataset. TableController now has a cbDataset that is added to whenever a new dataset is made by the user. Action listener on the dataset choicebox is only enabled once the user has imported.